renaming main test directory to make more consistent
