+++ v3/integration/lints/testdata/notAGolangFile.sh
